+++ packages/use-arcade/src/hooks/useLastChannelMessage.ts
@@ -1,0 +1,7 @@
+import { Message } from '../store'
+import { useChannelMessages } from './useChannelMessages'
+
+export const useLastChannelMessage: (channelId: string) => Message = (channelId: string) => {
+  const messages = useChannelMessages(channelId)
+  return messages[messages.length - 1]
+}
