+++ packages/use-arcade/src/index.ts
@@ -3,13 +3,14 @@ export { ArcadeContext } from './context'
 export { useArcadeRelay, UseArcadeRelayActions } from './hooks/useArcadeRelay'
 export { useBoolean } from './hooks/useBoolean'
 
+export { useChannelMessages, useLastChannelMessage } from './hooks'
+
 export {
   Channel,
   Message,
   setActiveChannelId,
   store,
   useActiveChannelId,
-  useChannelMessages,
   useChannelsCreated,
   useRideRequests,
 } from './store'
